+++ src/index.js
@@ -1,5 +1,7 @@
 import BaseRepository from './base-repository';
 import createController from './create-controller';
 import * as constants from './constants';
+import registerRoutes from './register-routes';
+import timePlugin from './time-plugin';
 
-export {BaseRepository, constants, createController};
+export {BaseRepository, constants, createController, registerRoutes, timePlugin};